+++ src/monitors/index.js
@@ -1,3 +1,4 @@
 module.exports = [
 	require('./waitingpods'),
+	require('./longnotready'),
 ];